FUSE: Associcate credentials to session leader
This helps a lot in case the environment variables of a shell
have changed after process creation. It is now sufficient to
run an external command (ie ls) on the fuse mount for the
new credentials to be picked up and associated to the shell.

Results in the following strange behavior, but not sure what
is the alternative, short of peeking at the PID's memory..

$ export KRB5CCNAME=FILE:/tmp/myticket
$ cd /eos/user # runs under bash PID, eosd unable to pick up credentials
bash: cd: /eos/user: Permission denied
$ ls /eos/user # new process, eosd associates creds to shell
0  a  b  c  d  e  f  g  h  i  j  k  l  m  n  o  p  proc  q  r  s  t  u  v  w  x  y  z
$ cd /eos/user #now works!