FUSE: Make XrdCl use fsuid of user when reading creds
My new patch adds support in XrdCl for options
xrd.secuid and xrd.secgid. The thread which reads
the credentials will temporarily set its fsuid
and fsgid to the values passed - this guarantees
that the user has permission to read those credentials,
and cannot play games with race conditions and symbolic links.