XrdFstOfsFile::close() now checks mIsInjection wherever it checks isReplication
The XrdFstOfsFile::close() method is not treating tape replicas
as "fully fledged", "first class citizen", file replicas.  One
very, very serious side effect of this is the
XrdFstOfsFile::close() method will delete an entire EOS file entry
from the EOS namespace (full data loss) when a retrieve from tape
is on-going, the tape server crashes and the file being written to
EOS disk never gets properly closed by the now dead tape server.
More specifically the following code contributes to what is to be
done when a file has not been properly closed:

1113       if (layOut->IsEntryServer() && (!isReplication)) {
1114         capOpaqueString += "&mgm.dropall=1";
1115       }

This code will send "&mgm.dropall=1" to the MGM even if there is
a tape replica of the file in question.  The code has now been
changed to the following:

1113       if (layOut->IsEntryServer() && (!isReplication) && (!mIsInjection)) {
1114         capOpaqueString += "&mgm.dropall=1";
1115       }

In fact all tests against !isReplication have been logically
"anded" with an additional test against !mIsInjection.