MGM: Avoid deadlock between the IConfigEngine mMutex and the FsView.ViewMutex. Fixes EOS-2692
This happened due to the fact that when the configuration was loaded and applied the mMutex
was first locked and then the ViewMutex. On the other hand when a file system configuration
was applied (possibly by some other thread) then things would go the other way around.
Now we force the order ViewMutex->mMutex also during the loading phase when booting.