[WFE] Adds a Prepare timestamp xattr
The WFE only forwards Prepare requests to CTA if there is not already a
Prepare request in-flight. This attribute records the time of the last
forwarded request.

This can be used by an operator to determine if the system has somehow
lost the Prepare request. If so, they can manually clear the Request ID
list so the request can be reissued by the user.