mgm: GroupBalancer: use sets for over/under avg group info
Since we don't seem to use the ptr to Group for almost all cases, just store the
names only. Since it is easy to convert the name to a FsView::Group just convert
this at the callsite. This should make the core of GroupBalancer easier to
separate from the part that deals with MGM/FsView.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>