MGM: groupbalancer: Allow GroupsInfoFetcher to take a filter fn
This allows for the group info fetcher to filter based on groups status, the
initial implementation uses a base functor of type GroupStatusFilter which any
new filter must implement. We will end up doing allocation, however since this class is
only ever going to be allocated at startup, it is not going to be a bottleneck.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>