COMMON: Introduce common::StringToNumeric
This function will convert a string to a numeric type, it uses std::from_chars
which promises no allocation, no throws. Since it is locale-independant as well
it is faster compared to functions like stol/f/d<> as well in many contexts.
However unfortunately we get floating point comparisions only in gcc11 onwards,
so for now a compile time check is applied to make sure that only integral types
are passed for gcc < 11. Tests are added to check the correct overloads chosen
and demonstrates some behaviour at edge cases. However currently the strings are
hard coded for tests which may fail in different compiler/arch settings
(std::to_string(numeric_limtits) might be portable way of checking
The function is added in the StringUtils header for now as this is less used
and reduces compilation overhead, and allows for specializations like adding
logging instead of strerrors in the future

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>