COMMON: Introduce an ObserverMgr class
This is a kind of lightweight event mgr like class that basically notifies
observers of a change, currently the callbacks itself are just a std::function
taking a variadic arg which must be specified at the time of construction.
Internally a map of slots and functions are stored, where the slot is supposed
to be supplied inorder to remove a callback. We hold a shared_ptr to a
std::function, and while invoking the function at the callsite we are only
given a list of weak_ptrs, this ensures deregistration without needing to block
at the callsite to hold a lock.

Fixes: EOS-5213
Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>