COMMON: ObserverMgr/CallbackList expose tag types
Since we need the observer tag type at call sites, expose this outside of the
class, Since both the classes are variadic right now, it must be invoked as
ObserverMgr<int,bool>::.., since we don't really depend on the templated types
expose this outside of the class. Also mark non copyable types as mutable.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>