MGM: GroupDrainer: handle retries and failed drains
This adds functionality to track the list of failed files on a drain and
retry (default: 5) attempts after a given interval. After this the FSes are
marked in a failed drain status and no longer tried.

Common drain status setting functionality is moved to a utils header as this
is not really directly concerned with GroupDrainer as such and maybe refactored
to reuse by drainer etc. at a later point.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>