MGM: Policy: avoid double map lookups & use constants where possible
Avoid double lookups where we use the map's value later.

Also use constants in code, this avoids potentially constructing a string while
doing a map lookup, while C++17 offers transparent lookups where keys are
guaranteed not to be constructed, this requires the map itself to be defined
with a std::less<> which requires changes in the API, however we can still
already construct these strings beforehand.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>