MGM: Introduce a cluster map type tracking the cluster status
This is a structure holding a vector of shared_ptrs holding the cluster data.
It is expected that call sites do not modify the shape of the data and only its
contents! We have an epoch to indicate when the clustermap changes shape, and
since we give out a shared_ptr to the data, when new disks/storage elements are
added the cluster map can change without affecting placements in progress

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>