MGM: placement: introduce a simple RoundRobin seed generator
This is just a vector of atomic counters, since we'd need to consistently Round
Robin in 2 dimensions when you'd want to iteratively select a group as well as
select disks, when one selects disks within groups we'd want the next one to be
taken remembering the previous selection, hence we'd need to remember what was
last chosen.

A simple atomic wrapper is used to make a vector of atomics as vector requires
copyable types for initialization/push_back ops. However we don't provide any
means of resizing the vector itself, so this class should be thread safe except
for initialization

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>