COMMON: Concurrency: Introduce a non blocking way to GC atomic ptrs
We introduce a simple Kernel RCU like primitive to track readers in their
critical sections. The idea is for rarely changing data, that readers
do not wait upon writers. The readers track when they enter and leave critical
sections and the writer side does not delete the old data until we're sure that
no readers access this.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>