COMMON: concurrency: introduce RAII like lock wrappers
Since some of the locks we have are stateful, we can't defer to sticking to just
shared_lock/unlock like functions to use the std::shared/scoped_lock. Wrappers
are added for Read and Write Locks. Currently we default to single writer mode
which can be changed in the future.

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>