COMMON: Concurrency: Move SimpleEpochCounter to experimental namespace
Due to collisions with hashing thread ids onto an array, we currently cannot
ensure the correct behaviour when trying to track per thread state. A future
TODO is to do this via thread local counters which should be performant and only
retrieve during RCUWriteLock, however this likely needs something
pthread_specific and also know when threads get destroyed to do safely.
Currently marking this experimental and for future work

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>