COMMON: Introduce an atomic unique ptr type
This is mainly only useful for data that is rarely changed, where you'd expect a
much better performance than a RWLock if the design allows for it. All
operations except the construction are threadsafe, reset does not delete the old
ptr but in turn returns it. So it is up to the callee to decide a safe point to
delete and reclaim memory when changing pointers. Tests are added to demonstrate
the various ThreadSafety features, SharedPtr resets are not threadsafe and is
shown with a failing test

Signed-off-by: Abhishek Lekshmanan <abhishek.lekshmanan@cern.ch>